Moved SlideUpdate to a new file and indentation.
