+++ lib/slide_update.dart
@@ -1,0 +1,15 @@
+import 'package:liquid_swipe/Constants/Helpers.dart';
+
+class SlideUpdate {
+  final UpdateType updateType;
+  final SlideDirection direction;
+  final double slidePercentHor;
+  final double slidePercentVer;
+
+  SlideUpdate(
+    this.direction,
+    this.slidePercentHor,
+    this.slidePercentVer,
+    this.updateType,
+  );
+}
